Split helper functions in multiple files
