+++ pkg/helpers/version.go
@@ -7,37 +7,8 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/fatih/color"
 	"github.com/hashicorp/go-version"
 )
-
-// ColoredVersion returns a colored version string
-func ColoredVersion(version string) string {
-	if IsPreReleaseVersion(version) {
-		return color.YellowString(version)
-	}
-	return color.CyanString(version)
-}
-
-// ColoredInstallHelper returns a colored install helper string
-func ColoredInstallHelper(version string) string {
-	return color.CyanString(fmt.Sprintf("`tfversion install %s`", version))
-}
-
-func ColoredListHelper() string {
-	return color.CyanString("`tfversion list`")
-}
-
-// ExitWithError prints an error message and exits with status code 1
-func ExitWithError(message string, err error) {
-	fmt.Printf("%s %s: %s\n", color.HiRedString("error:"), message, err)
-	os.Exit(1)
-}
-
-// ErrorWithHelp returns an error with a help message
-func ErrorWithHelp(help string) error {
-	return fmt.Errorf("see %s for help and examples", color.CyanString("`%s`", help))
-}
 
 // IsPreReleaseVersion checks if the given version is a Terraform pre-release version
 func IsPreReleaseVersion(version string) bool {
